Move glTF writing into ConvertCommand itself
Move writing the glTF 2.0 file (JSON) and its binary buffer file
(*.glbuf) into the ConvertCommand itself, away from the GeometryCache.
